+++ src/contexts/MapContext.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React, { createContext, useState, useContext } from 'react';
 import { SchoolMapPin } from '@/types/school';
